requirements.txt
geopandas==0.10.0
streamlit==0.89.0
folium==0.12.1
pandas==1.2.4
plotly==5.3.1
streamlit_folium==0.4.0
Pillow==8.3.2
